chore: reorganize auth code
